Refactor to move ResourceID to swapi package.
