+++ resolver/starwars/person.go
@@ -7,6 +7,7 @@ import (
 
 	"gqlgen-starwars/errors"
 	"gqlgen-starwars/loaders"
+	swapi2 "gqlgen-starwars/swapi"
 	"gqlgen-starwars/utils"
 )
 
@@ -27,7 +28,7 @@ func (r *personResolver) Films(ctx context.Context, p *swapi.Person) ([]*swapi.F
 	ids := make([]int, 0, len(p.FilmURLs))
 
 	for _, url := range p.FilmURLs {
-		id, err := utils.ResourceId(string(url))
+		id, err := swapi2.ResourceId(string(url))
 		if err != nil {
 			entry.WithError(err).Error("Failed to parse id from url")
