+++ swapi/swapi.go
@@ -1,4 +1,4 @@
-package utils
+package swapi
 
 import (
 	"strconv"
